requirements.txt
crewai>=0.11.0
playwright>=1.41.0
python-dotenv>=1.0.0
langchain>=0.1.0
pandas>=2.0.0
requests>=2.31.0 